+++ Model/Google/Symbol.cs
@@ -5,6 +5,6 @@
         public Property Property { get; set; }
         public BoundingBox BoundingBox { get; set; }
         public string Text { get; set; }
-        public double Confidence { get; set; }
+        public decimal Confidence { get; set; }
     }
 }
