+++ Model/Google/WebEntity.cs
@@ -3,7 +3,7 @@
     public class WebEntity : ILabel
     {
         public string EntityId { get; set; }
-        public double Score { get; set; }
+        public decimal Score { get; set; }
         public string Description { get; set; }
     }
 }
