+++ Model/Microsoft/Caption.cs
@@ -3,6 +3,6 @@
     public class Caption
     {
         public string Text { get; set; }
-        public double Confidence { get; set; }
+        public decimal Confidence { get; set; }
     }
 }
