+++ Model/Microsoft/Hair.cs
@@ -4,7 +4,7 @@ namespace Model.Microsoft
 {
     public class Hair
     {
-        public double Bald { get; set; }
+        public decimal Bald { get; set; }
         public bool Invisible { get; set; }
         public IList<HairColor> HairColor { get; set; }
     }
