+++ Model/Microsoft/ImageObject.cs
@@ -4,6 +4,6 @@
     {
         public Rectangle Rectangle { get; set; }
         public string Object { get; set; }
-        public double Confidence { get; set; }
+        public decimal Confidence { get; set; }
     }
 }
